try g

+++ printf/utilites_for_double/ft_handle_g.c
@@ -28,7 +28,7 @@ char 	*ft_check_g(int exponenta, char *string, t_arg *func, long double nbr)
 
 	acc = func->acc;
 
-	len = (int)ft_strlen(ft_itoa((int)nbr));
+	len = (long int)ft_strlen(ft_itoa((long int)nbr));
 	func->acc = func->acc - len;
 	func->acc = func->acc >= 0 ? func->acc : 0;
